Move test-related code to own subdir
