+++ src/test/test-harness.c
@@ -29,7 +29,7 @@
  */
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #include <stdarg.h>
 #include <stdlib.h>
